#38 refactor: entity modification
