+++ src/main/java/kr/oshino/eataku/restaurant/admin/entity/HashTag.java
@@ -1,0 +1,6 @@
+package kr.oshino.eataku.restaurant.admin.entity;
+
+public enum HashTag {
+
+    맛있어요, 친절해요, 쾌적해요, 다시갈래요, 분위기가_좋아요;
+}
